pyproject.toml
[project]
name = "mqtt-doorbell"
version = "0.0.9"
dependencies = ["paho-mqtt"]
authors = [
  { name="Alex Bratchik", email="abratchik@gmail.com" },
]
description = "A simple Linux-based doorbell, which plays the mp3 bell ringtone upon receiving an MQTT message."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.urls]
Homepage = "https://github.com/abratchik/mqtt-doorbell"
Issues = "https://github.com/abratchik/mqtt-doorbell/issues"